added order

+++ exercise-2/app/models/Order.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.Json
 
-case class Order(var id: Float, var basketId: Float, var paymentId: Float)
+case class Order(var id: Long, var basketId: Long, var paymentId: Long)
 
 object Order{
   implicit val orderFormat = Json.format[Order];
